pc/requirements.txt
appdirs==1.4.0
argcomplete==1.4.1
numpy==1.11.1
opencv-python==3.1.0.1
Pillow==3.3.1
pyserial==3.1.1
python-aalib==0.3.2
python-escpos==2.2.0
pyusb==1.0.0
PyYAML==3.12
qrcode==5.3
six==1.10.0
